storage: remove legacy SST iterator
The legacy sstable iterator is no longer in use. Removing the legacy
sstable iterator now also avoids updating it to accommodate the new
Pebble internal iterator interface with lazy value fetching.

Release note: None
