+++ pkg/storage/BUILD.bazel
@@ -37,7 +37,6 @@ go_library(
         "slice.go",
         "slice_go1.9.go",
         "sst.go",
-        "sst_iterator.go",
         "sst_writer.go",
         "store_properties.go",
         "temp_engine.go",
@@ -129,7 +128,6 @@ go_test(
         "pebble_test.go",
         "read_as_of_iterator_test.go",
         "resource_limiter_test.go",
-        "sst_iterator_test.go",
         "sst_test.go",
         "sst_writer_test.go",
         "temp_engine_test.go",
