+++ pkg/storage/bench_pebble_test.go
@@ -537,13 +537,9 @@ func BenchmarkSSTIterator(b *testing.B) {
 	defer log.Scope(b).Close(b)
 	for _, numKeys := range []int{1, 100, 10000} {
 		b.Run(fmt.Sprintf("keys=%d", numKeys), func(b *testing.B) {
-			for _, variant := range []string{"legacy", "pebble"} {
-				b.Run(fmt.Sprintf("variant=%s", variant), func(b *testing.B) {
-					for _, verify := range []bool{false, true} {
-						b.Run(fmt.Sprintf("verify=%t", verify), func(b *testing.B) {
-							runSSTIterator(b, variant, numKeys, verify)
-						})
-					}
+			for _, verify := range []bool{false, true} {
+				b.Run(fmt.Sprintf("verify=%t", verify), func(b *testing.B) {
+					runSSTIterator(b, numKeys, verify)
 				})
 			}
 		})
