+++ pkg/storage/bench_test.go
@@ -1749,7 +1749,7 @@ func runCheckSSTConflicts(
 	}
 }
 
-func runSSTIterator(b *testing.B, variant string, numKeys int, verify bool) {
+func runSSTIterator(b *testing.B, numKeys int, verify bool) {
 	keyBuf := append(make([]byte, 0, 64), []byte("key-")...)
 	value := MVCCValue{Value: roachpb.MakeValueFromBytes(bytes.Repeat([]byte("a"), 128))}
 
@@ -1765,27 +1765,13 @@ func runSSTIterator(b *testing.B, variant string, numKeys int, verify bool) {
 	}
 	sstWriter.Close()
 
-	var makeSSTIterator func(data []byte, verify bool) (SimpleMVCCIterator, error)
-	switch variant {
-	case "legacy":
-		makeSSTIterator = func(data []byte, verify bool) (SimpleMVCCIterator, error) {
-			return NewLegacyMemSSTIterator(data, verify)
-		}
-	case "pebble":
-		makeSSTIterator = func(data []byte, verify bool) (SimpleMVCCIterator, error) {
-			return NewMemSSTIterator(data, verify, IterOptions{
-				KeyTypes:   IterKeyTypePointsAndRanges,
-				LowerBound: keys.MinKey,
-				UpperBound: keys.MaxKey,
-			})
-		}
-	default:
-		b.Fatalf("unknown variant %q", variant)
-	}
-
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		iter, err := makeSSTIterator(sstFile.Bytes(), verify)
+		iter, err := NewMemSSTIterator(sstFile.Bytes(), verify, IterOptions{
+			KeyTypes:   IterKeyTypePointsAndRanges,
+			LowerBound: keys.MinKey,
+			UpperBound: keys.MaxKey,
+		})
 		if err != nil {
 			b.Fatal(err)
 		}
